requirements.txt
python-dotenv
requests
garth
fit_tool
garminconnect
